backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.52.2
python-dotenv==1.0.1
requests==2.32.3


